Tried FIGMA Auto-CSS/HTML
I was hoping to get around the bummer that CSS and HTML are by relying on FIGMA's generation features. I have left evidence of my attempts commented out in the files, for group review.

Turns out, not only is it impossible to rely completely on the feature at the best of times, but due to my awful design skills, the styling is mostly useless to us, with half-generated nonsense spilling everywhere.

This is going to add a lot of time to completion. You lot will need to chip in.