Course Prerequisites, Made Pages
Added course prerequisites to the db structure. Another M-t-M, similar issues to the friends system.

I've instantiated some basic pages for our primary URLs, so that you don't get signed out when you press the nav buttons. The only content I've added is the start of the friends page: when referencing an object's database relations, remember to re-pull the object from the DB, or the set will not be included.

Also, Henry, you did not fix that fucking typo. That was literally half of your last commit, and "account" was still missing the 'n'.