Courses in-progress updates
Going to work on the presentation for a while, now. Will come back to finish this off. Incomplete, but I have an ok idea of how to finish the rest. Hopefully done by 5ish this evening.

The "testing" part of the rubric has me worried. That will probably take some time.